backend/requirements.txt
Django~=5.1.6
djangorestframework~=3.15.2
drf-yasg~=1.21.9
djangorestframework_simplejwt~=5.4.0
social-auth-app-django
pillow